feat(db): add db connection in the application
Created connection to the database and registered it into the application data (app_data). This connection gets stored as part of the application state as an Arc (smart pointer that creates references to it across multiple instances of the application), which then can be used wherever necessary, using the Data extractor to extract information of the app_data.